[feature] Optimize the code
