+++ swama/Sources/SwamaKit/Model/ModelDownloader.swift
@@ -147,7 +147,7 @@ public enum ModelDownloader {
 
                 // Download with resume support
                 printMessage("[\(fileIndex)/\(totalFiles)] Downloading: \(fileInfo.fileName)")
-                try await downloadWhisperKitFileWithResume(
+                try await downloader.downloadWhisperKitFileWithResume(
                     from: fileInfo.url,
                     to: fileInfo.localPath,
                     totalSize: remoteSize,
@@ -222,175 +222,6 @@ public enum ModelDownloader {
             }
         }
         return total
-    }
-
-    static func downloadWhisperKitFileWithResume(
-        from urlString: String,
-        to localFile: URL,
-        totalSize: Int64,
-        localSize: Int64
-    ) async throws {
-        guard let url = URL(string: urlString) else {
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 14,
-                userInfo: [NSLocalizedDescriptionKey: "Invalid URL: \(urlString)"]
-            )
-        }
-
-        var request = URLRequest(url: url)
-        var isResuming = false
-        if localSize > 0, localSize < totalSize {
-            request.setValue("bytes=\(localSize)-", forHTTPHeaderField: "Range")
-            isResuming = true
-        }
-
-        let (byteStream, response) = try await URLSession.shared.bytes(for: request)
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 9,
-                userInfo: [NSLocalizedDescriptionKey: "Invalid response from server."]
-            )
-        }
-
-        let statusCode = httpResponse.statusCode
-        if !(200 ... 299).contains(statusCode) {
-            if statusCode == 404 {
-                throw NSError(
-                    domain: "ModelDownloader",
-                    code: 404,
-                    userInfo: [NSLocalizedDescriptionKey: "HTTP 404"]
-                )
-            }
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 15,
-                userInfo: [NSLocalizedDescriptionKey: "HTTP error \(statusCode)"]
-            )
-        }
-
-        let shouldResume = isResuming && statusCode == 206 // 206 Partial Content
-
-        if !FileManager.default.fileExists(atPath: localFile.path) {
-            FileManager.default.createFile(atPath: localFile.path, contents: nil)
-        }
-
-        let handle = try FileHandle(forWritingTo: localFile)
-        defer { try? handle.close() }
-
-        if shouldResume {
-            try handle.seekToEnd()
-        }
-        else {
-            try handle.truncate(atOffset: 0)
-        }
-
-        let progressBar = ProgressBar(total: totalSize, initial: shouldResume ? localSize : 0)
-        var buffer: [UInt8] = []
-
-        for try await byte in byteStream {
-            buffer.append(byte)
-            if buffer.count >= 8192 {
-                try handle.write(contentsOf: Data(buffer))
-                progressBar.update(bytes: Int64(buffer.count))
-                buffer.removeAll(keepingCapacity: true)
-            }
-        }
-        if !buffer.isEmpty {
-            try handle.write(contentsOf: Data(buffer))
-            progressBar.update(bytes: Int64(buffer.count))
-        }
-
-        progressBar.finish()
-    }
-
-    static func downloadWhisperKitFile(
-        from urlString: String,
-        to localFile: URL,
-        description: String,
-        fileIndex: Int,
-        totalFiles: Int
-    ) async throws {
-        // Check if file already exists and is not empty
-        if FileManager.default.fileExists(atPath: localFile.path) {
-            do {
-                let attributes = try FileManager.default.attributesOfItem(atPath: localFile.path)
-                if let fileSize = attributes[.size] as? Int64, fileSize > 0 {
-                    printMessage("[\(fileIndex)/\(totalFiles)] Exists: \(description), skip")
-                    return
-                }
-            }
-            catch {
-                // If we can't get attributes, assume file is corrupt and re-download
-            }
-        }
-
-        guard let downloadURL = URL(string: urlString) else {
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 14,
-                userInfo: [NSLocalizedDescriptionKey: "Invalid URL: \(urlString)"]
-            )
-        }
-
-        printMessage("[\(fileIndex)/\(totalFiles)] Downloading: \(description)")
-
-        // Use the same download logic as regular models
-        let (byteStream, response) = try await URLSession.shared.bytes(from: downloadURL)
-
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 9,
-                userInfo: [NSLocalizedDescriptionKey: "Invalid response from server for file \(description)."]
-            )
-        }
-
-        if httpResponse.statusCode == 404 {
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 404,
-                userInfo: [NSLocalizedDescriptionKey: "HTTP 404 for \(description)"]
-            )
-        }
-        else if !(200 ... 299).contains(httpResponse.statusCode) {
-            throw NSError(
-                domain: "ModelDownloader",
-                code: 15,
-                userInfo: [NSLocalizedDescriptionKey: "HTTP \(httpResponse.statusCode) for \(description)"]
-            )
-        }
-
-        // Get expected content length for progress bar
-        let expectedLength = httpResponse.expectedContentLength > 0 ? httpResponse.expectedContentLength : 1024 * 1024
-        let progressBar = ProgressBar(total: expectedLength)
-
-        // Create file for writing
-        if !FileManager.default.fileExists(atPath: localFile.path) {
-            FileManager.default.createFile(atPath: localFile.path, contents: nil)
-        }
-
-        let handle = try FileHandle(forWritingTo: localFile)
-        defer { try? handle.close() }
-
-        try handle.truncate(atOffset: 0)
-
-        var buffer: [UInt8] = []
-        for try await byte in byteStream {
-            buffer.append(byte)
-            if buffer.count >= 8192 {
-                try handle.write(contentsOf: Data(buffer))
-                progressBar.update(bytes: Int64(buffer.count))
-                buffer.removeAll(keepingCapacity: true)
-            }
-        }
-        if !buffer.isEmpty {
-            try handle.write(contentsOf: Data(buffer))
-            progressBar.update(bytes: Int64(buffer.count))
-        }
-
-        progressBar.finish()
     }
 
     // MARK: - WhisperKit Helper Functions
